telegrambot/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
certifi==2019.11.28       # via python-telegram-bot
cffi==1.14.0              # via cryptography
click==7.1.1              # via -r requirements.in, flask, pip-tools
cryptography==2.8         # via python-telegram-bot
decorator==4.4.2          # via python-telegram-bot
flask-sqlalchemy==2.4.1   # via -r requirements.in
flask==1.1.1              # via -r requirements.in, flask-sqlalchemy
future==0.18.2            # via python-telegram-bot
itsdangerous==1.1.0       # via flask
jinja2==2.11.1            # via flask
markupsafe==1.1.1         # via jinja2
pip-tools==4.5.1          # via -r requirements.in
pycparser==2.20           # via cffi
python-dotenv==0.12.0     # via -r requirements.in
python-telegram-bot==12.4.2  # via -r requirements.in
six==1.14.0               # via cryptography, pip-tools
sqlalchemy==1.3.15        # via -r requirements.in, flask-sqlalchemy
tornado==6.0.4            # via python-telegram-bot
werkzeug==1.0.0           # via flask
